[NEW] Adds a new view helper for links
This is necessary because with the next commit external packages can deliver the content.
This view helper is not much more than the default link.action view helper but it set the package always to Scriptme.MarkdownContent. It's basically a convenience feature as you could do the same thing by setting the "package" property in the default link.action view helper