[NEW] Includes basic version of include view helper
This add the include view helper. This can be used to render a specific package->controller->action in your template. Think of it as plugin include. This is more or less a copy of the Neos Plugin implementation. While it is basically working, the action won't receive any arguments and link will be generated incorrectly. There are also some problems with the routing to solve.