[FEATURE] Cleanup of the Site Package handling
Use the default view of the controller and just override the template path.
To make it easier to work with the md viewhelper, a Context was created that will
hold the current SitePackageKey and other important information.
With this Context we don't need to create the resource and content path every time
only if we override the package