[NEW] Adds a new option "addRoot" to menu view helper
This option will add the root of the menu (or in other words it's starting point) on top of the resulting pages structure.
So if you have something link

page1
	page 1.1
	page 1.2

And you start your menu at "page1" the result page list will look like [page1, page1.1, page1.2] if the option "addRoot" is set. If not, "page1" would not be part of the result.