[FEATURE] Configure a custom 404 page
At the moment the user would just get a ugly 404 error, if
the the requested page dosen't exists. This patch makes it
possible to configure a fallback page that should be used
and only display the ugly 404 error if the fallback page is not
configured or dosen't exists either.

Scriptme:
  MarkdownContent:
    errorHandling:
      404: '404'